requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.14.6
Shapely==2.0.3 